Split align related css property parsing out into its own file https://bugs.webkit.org/show_bug.cgi?id=279299
Reviewed by Darin Adler.

Splits property consumers for CSS Align properties into their own file and refactor
them a bit to hide the implementation details.

We now have explicit functions for each property, named appropriately to match the
property. The implementations remain the same, still calling to shared consumers
and passing predicates to refine behavior, but from the outside, things are much
more clear.

The shorthands were also refactored and now all use a shared consumeAlignShorthand
consumer function.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/parser/CSSPropertyParser.cpp:
* Source/WebCore/css/parser/CSSPropertyParser.h:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Align.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Align.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
* Source/WebCore/css/process-css-properties.py:

Canonical link: https://commits.webkit.org/283317@main
